warns: NewFeature: You can select action when max warn exceeded

+++ sophie_bot/modules/warns.py
@@ -23,6 +23,7 @@ from aiogram.types.inline_keyboard import (
     InlineKeyboardButton,
     InlineKeyboardMarkup
 )
+from babel.dates import format_timedelta
 from bson.objectid import ObjectId
 
 from sophie_bot import BOT_ID
@@ -30,7 +31,8 @@ from sophie_bot.decorator import register
 from sophie_bot.services.mongo import db
 from .utils.connections import chat_connection
 from .utils.language import get_strings_dec
-from .utils.restrictions import ban_user
+from .utils.message import convert_time, InvalidTimeUnit
+from .utils.restrictions import ban_user, mute_user
 from .utils.user_details import (
     get_user_and_text_dec, get_user_dec,
     get_user_link, is_user_admin
@@ -91,10 +93,16 @@ async def warn(message, chat, user, text, strings):
         max_warn = 3
 
     if warns_count >= max_warn:
-        await ban_user(str(chat_id), str(user_id))
-        await message.reply(strings['warn_bun'].format(user=member))
-        db.warns_v2.delete_many({'user_id': user_id, 'chat_id': chat_id})
-        return
+        if await max_warn_func(chat_id, user_id):
+            await db.warns_v2.delete_many({'user_id': user_id, 'chat_id': chat_id})
+            data = await db.warnmode.find_one({'chat_id': chat_id})
+            if data['mode'] != 'tmute':
+                return await message.reply(strings['max_warn_exceeded'] % (member,
+                                           strings['banned'] if data['mode'] == 'ban' else strings['muted']))
+            else:
+                return await message.reply(strings['max_warn_exceeded:tmute'] % (member,
+                                           format_timedelta(convert_time(data['time']),
+                                                            locale=strings['language_info']['babel'])))
     else:
         text += strings['warn_num'].format(curr_warns=warns_count, max_warns=max_warn)
 
@@ -162,10 +170,10 @@ async def warnlimit(message, chat, strings):
     elif not arg[0].isdigit():
         return await message.reply(strings['not_digit'])
     else:
-        if arg[0] < 2:
+        if int(arg[0]) < 2:
             return await message.reply(strings['warnlimit_short'])
 
-        elif arg[0] > 10000:  # Max value
+        elif int(arg[0]) > 10000:  # Max value
             return await message.reply(strings['warnlimit_long'])
 
         new = {
@@ -174,7 +182,7 @@ async def warnlimit(message, chat, strings):
         }
 
         await db.warnlimit.update_one({'chat_id': chat_id}, {'$set': new}, upsert=True)
-        await message.reply(strings['warnlimit_updated'].format(num=arg[0]))
+        await message.reply(strings['warnlimit_updated'].format(num=int(arg[0])))
 
 
 @register(cmds=['resetwarns', 'delwarns'], user_can_restrict_members=True)
@@ -201,13 +209,85 @@ async def reset_warn(message, chat, user, strings):
         await message.reply(strings['usr_no_wrn'].format(user=user))
 
 
+@register(cmds='warnmode', user_admin=True, bot_can_restrict_members=True)
+@chat_connection(admin=True)
+@get_strings_dec('warns')
+async def warnmode(message, chat, strings):
+    chat_id = chat['chat_id']
+    acceptable_args = ['ban', 'tmute', 'mute']
+    arg = str(message.get_args()).split()
+    new = {'chat_id': chat_id}
+
+    if arg and arg[0] in acceptable_args:
+        option = ''.join(arg[0])
+        if (data := await db.warnmode.find_one({'chat_id': chat_id})) is not None and data['mode'] == option:
+            return await message.reply(strings['same_mode'])
+        if arg[0] == acceptable_args[0]:
+            new['mode'] = option
+            await db.warnmode.update_one({'chat_id': chat_id},
+                                         {'$set': new}, upsert=True)
+        elif arg[0] == acceptable_args[1]:
+            try:
+                time = arg[1]
+            except IndexError:
+                return await message.reply(strings['no_time'])
+            else:
+                try:
+                    # For better UX we have to show until time of tmute when action is done.
+                    # We can't store timedelta class in mongodb; Here we check validity of given time.
+                    convert_time(time)
+                except (InvalidTimeUnit, TypeError):
+                    return await message.reply(strings['invalid_time'])
+                else:
+                    new['mode'] = option
+                    new['time'] = time
+                    await db.warnmode.update_one({'chat_id': chat_id},
+                                                 {'$set': new}, upsert=True)
+        elif arg[0] == acceptable_args[2]:
+            new['mode'] = option
+            await db.warnmode.update_one({'chat_id': chat_id},
+                                         {'$set': new}, upsert=True)
+        await message.reply(strings['warnmode_success'] % (chat['chat_title'], option))
+    else:
+        text = strings['wrng_args'] % '\n'.join(acceptable_args)
+        if (curr_mode := await db.warnmode.find_one({'chat_id': chat_id})) is not None:
+            mode = curr_mode['mode']
+            text += strings['mode_info'] % mode
+        await message.reply(text)
+
+
+async def max_warn_func(chat_id, user_id):
+    if mode := await db.warnmode.find_one({'chat_id': chat_id}):
+        if mode['mode'] == 'ban':
+            return await ban_user(chat_id, user_id)
+        elif mode['mode'] == 'tmute':
+            return await mute_user(chat_id, user_id, convert_time(mode['time']))
+        elif mode['mode'] == 'mute':
+            return await mute_user(chat_id, user_id)
+    else:  # Default
+        return ban_user(chat_id, user_id)
+
+
 async def __export__(chat_id):
     if data := await db.warnlimit.find_one({'chat_id': chat_id}):
         number = data['num']
     else:
         number = 3
 
-    return {'warns': {'warns_limit': number}}
+    if data := await db.warnmode.find_one({'chat_id': chat_id}):
+        if (mode := data['mode']) == 'tmute':
+            new = {
+                'mode': mode,
+                'time': data['time']
+            }
+        else:
+            new = {
+                'mode': mode
+            }
+    else:
+        new = None  # null
+
+    return {'warns': {'warns_limit': number, 'warn_mode': new}}
 
 
 async def __import__(chat_id, data):
@@ -220,3 +300,15 @@ async def __import__(chat_id, data):
             return
 
         await db.warnlimit.update_one({'chat_id': chat_id}, {'$set': {'num': number}}, upsert=True)
+
+    if (mode := data['warn_mode']) is not None:
+        if mode['mode'] == 'tmute':
+            new = {
+                'mode': mode['mode'],
+                'time': mode['time']
+            }
+        else:
+            new = {
+                'mode': mode['mode']
+            }
+        await db.warnmode.update_one({'chat_id': chat_id}, {'$set': new}, upsert=True)
